Create schedule_J_cleaner.R
This function "cleans" the data insofar as it takes the information on the first person listed on the schedule J form (usually the CEO or highest paid individual). This function was meant to prepare the data for an s3 bucket in AWS so the data could be queries using Athena.